requirements.txt
ultralytics==8.3.225
huggingface_hub==1.1.2
opencv-python==4.12.0.88
matplotlib==3.10.7
numpy==2.2.6
fastapi==0.121.0
uvicorn==0.38.0
python-multipart==0.0.20
pydantic==2.12.4
langchain==1.0.4
langchain-openai==1.0.2
openai==2.7.1
python-dotenv==1.2.1